Insert generic import.py
